+++ src/library/DIPS.Mobile.UI/Components/Sorting/SortControl.Properties.cs
@@ -25,6 +25,20 @@ public partial class SortControl
             OnSelectedItemChanged();   
         }
     }
+    
+    public static readonly BindableProperty HasHapticsProperty = BindableProperty.Create(
+        nameof(HasHaptics),
+        typeof(bool),
+        typeof(SortControl), defaultValue:true);
+
+    /// <summary>
+    /// Determines if the phone should stimulate the sense of touch and motion by the use of vibration when people select sort options.
+    /// </summary>
+    public bool HasHaptics
+    {
+        get => (bool)GetValue(HasHapticsProperty);
+        set => SetValue(HasHapticsProperty, value);
+    }
 
     public object? InitialSelectedItem
     {